splitting into relevant file structure
