+++ package.yaml
@@ -4,6 +4,7 @@ version: 0.0.0
 dependencies:
   - ansi-terminal
   - base
+  - extra
   - containers
   - inspection-testing
   - mtl
